Fix: extract root widget
